Add lambdas deployment
We will aliases for lambda function. We will have 2 aliases - production and dev. The stages for API will be two as well and each stage will points to the specific lambda alias. lambda_package pack the lamda and sets the package property for same target in lambda_deploy(in current example lambda_package:all task will set lambda_deploy.all.package). We use same code for our lamda but different entry point for simplicity so we need to pack only once and deploy 2 lamda with different entrypoint(handler in the config).